# Removes ModuleInstance.TypeIDsIndex to reduce allocations