Ref #889 

This further reduces the allocation during instantiation:

```
$ benchstat old.txt new.txt                                                                                                                    
name                                    old time/op    new time/op    delta
Initialization/interpreter-10             23.7µs ±11%    22.7µs ± 8%    ~     (p=0.421 n=5+5)
Initialization/interpreter-multiple-10    10.8µs ± 7%    11.2µs ± 7%    ~     (p=0.548 n=5+5)
Initialization/compiler-10                33.1µs ± 2%    32.9µs ± 1%    ~     (p=0.905 n=5+4)
Initialization/compiler-multiple-10       15.0µs ± 9%    16.0µs ±16%    ~     (p=0.421 n=5+5)

name                                    old alloc/op   new alloc/op   delta
Initialization/interpreter-10              141kB ± 0%     141kB ± 0%    ~     (p=0.079 n=4+5)
Initialization/interpreter-multiple-10     143kB ± 0%     142kB ± 0%  -0.30%  (p=0.008 n=5+5)
Initialization/compiler-10                 145kB ± 0%     144kB ± 0%  -0.29%  (p=0.008 n=5+5)
Initialization/compiler-multiple-10        145kB ± 0%     145kB ± 0%  -0.29%  (p=0.016 n=5+4)

name                                    old allocs/op  new allocs/op  delta
Initialization/interpreter-10               85.0 ± 0%      83.0 ± 0%  -2.35%  (p=0.008 n=5+5)
Initialization/interpreter-multiple-10      92.0 ± 0%      90.0 ± 0%  -2.17%  (p=0.008 n=5+5)
Initialization/compiler-10                  35.0 ± 0%      33.0 ± 0%  -5.71%  (p=0.008 n=5+5)
Initialization/compiler-multiple-10         41.0 ± 0%      39.0 ± 0%  -4.88%  (p=0.008 n=5+5)


```